adding new class

+++ src/test/java/steps/Hooks.java
@@ -2,6 +2,7 @@ package steps;
 
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
+import io.cucumber.java.Scenario;
 import org.openqa.selenium.WebDriver;
 import utilities.Driver;
 
@@ -9,21 +10,26 @@ import java.util.concurrent.TimeUnit;
 
 public class Hooks {
 
-    WebDriver driver = Driver.gerDriver();
+    WebDriver driver;
 
     @Before
-    public void setUp(){
+    public void setUp(Scenario scenario) {
+        if (!scenario.getSourceTagNames().contains("@api")) {
 
-       driver.manage().window().maximize();
-       driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
-       driver.manage().deleteAllCookies();
-        System.out.println("Before Scenario Method");
+            driver = Driver.gerDriver();
+            driver.manage().window().maximize();
+            driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
+            driver.manage().deleteAllCookies();
+            System.out.println("Before Scenario Method");
+        }
     }
 
     @After
-    public void tearDown () throws InterruptedException {
-        Thread.sleep(2000);
-        driver.quit();
-        System.out.println("After Scenario Method");
+    public void tearDown(Scenario scenario) throws InterruptedException {
+        if (!scenario.getSourceTagNames().contains("@api")) {
+            Thread.sleep(3000);
+            driver.quit();
+            System.out.println("After Scenario Method");
+        }
     }
 }
